реализует подписки внутри компонентов и производит рефакторинг файла utils, переносит вспомогательные ф-ии из trip-info, event, new-event
